Add the final modification of simulator

+++ inc/jitter_info.h
@@ -2,11 +2,11 @@
 #define __JITTER_INFO_H
 
 typedef struct Jitter_constraint {
-	float str_jitter_bound; // The margin of start time jitter
-	float exe_jitter_bound; // The margin of execution time jitter
-	float fin_jitter_bound; // The margin of finishing time jitter
-	float alpha; // A percentage ranging from 0~100
-	float fin_time_target;
+	double str_jitter_bound; // The margin of start time jitter
+	double exe_jitter_bound; // The margin of execution time jitter
+	double fin_jitter_bound; // The margin of finishing time jitter
+	double alpha; // A percentage ranging from 0~100
+	double fin_time_target;
 } jitter_constraint;
 
 #endif // __JITTER_INFO_H
